requirements.txt
flask
requests
sqlalchemy-libsql
python-dotenv
